requirement.txt
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
attrs==21.2.0
autopep8==2.0.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
cisco-anyconnect-cli==0.5
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
fasteners==0.14.1
fs==2.4.16
future==0.18.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.2
httpx==0.23.3
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
keepasshttp==1.4.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
ordereddict==1.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pip-init==0.3.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.10.0
pycryptodome==3.15.0
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyTelegramBotAPI==4.9.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.21.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
systemd-python==234
telebot==0.0.4
texttable==1.6.4
tomli==2.0.1
tornado==6.2
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
websocket-client==1.2.3
xdg==5
xkit==0.0.0
zipp==1.0.0
